requirements.txt
# Core frameworks
langchain>=0.2.16
langgraph>=0.2.27
langchain-openai>=0.1.22
langchain-community>=0.2.11
openai>=1.51.0
langchain-text-splitters>=0.2.0

# Memory & Vector stores
chromadb>=0.5.5
faiss-cpu>=1.8.0
tiktoken>=0.7.0

# Tools
duckduckgo-search>=6.2.5
requests>=2.32.3
arxiv>=2.1.0
semanticscholar>=0.6.3
pypdf>=5.0.0
playwright>=1.47.0
fastapi>=0.112.0
uvicorn[standard]>=0.30.0
python-docx>=1.1.0
openpyxl>=3.1.0

# Utils
pydantic>=2.9.2
rich>=13.9.2
tenacity>=9.0.0